Don't use U mode, removed in python3.11
This patch was helpfully submitted by Bas Couwenberg to drop use of
the universal newline flag, since Python 3.11 no longer supports it.
It probably breaks the ability to build new correlation files under
Python 2.7 and earlier, but since it shouldn't affect operation of
the utility with prebuilt correlations (the way it's typically
used), this isn't yet considered to drop Python 2.7 support
altogether.