Implement time zone aware expiration comparisons
Instead of using a naive time comparison, check times relative to
the user's time zone (shifted to UTC) and the expiration times
embedded in NWS documents (implicitly UTC).

Drop the ugly hack of the 24-hour expiry delay, but keep it at 1
hour to still provide some protection against premature filtering in
case someone switches into or out of DST on a different date than
NWS expects it to happen.